requirements.txt
numpy==2.3.2
Pillow==11.3.0
torch==2.8.0
torchvision==0.23.0
